Changed project name and added Github Actions Squashed commits: [e2453c4] Test placeholder [e8500d8] "Tests" tweaks since Maven is bitching about it [5cc1b41] build.yml fix to use Java 17 [10647fc] Additional tweaks [d0747c7] Changed project name [a589d5b] Fixed build.yml [fe7c177] Updated build.yml steps with Sonar Cloud scenario
