+++ src/main/java/io/github/mityavasilyev/springreactbarapp/SpringReactBarAppApplication.java
@@ -1,0 +1,14 @@
+package io.github.mityavasilyev.springreactbarapp;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class SpringReactBarAppApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(SpringReactBarAppApplication.class, args);
+
+    }
+
+}
