+++ src/main/java/io/github/mityavasilyev/springreactbarapp/extra/Unit.java
@@ -1,0 +1,8 @@
+package io.github.mityavasilyev.springreactbarapp.extra;
+
+/**
+ * Provides unit type constants for measurements
+ */
+public enum Unit {
+    OUNCE, MILLILITER, LITER, GRAM, PIECE
+}
