--- src/main/java/io/github/mityavasilyev/springvertxreactbarapp/SpringVertxReactBarAppApplication.java
@@ -1,14 +1,0 @@
-package io.github.mityavasilyev.springvertxreactbarapp;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class SpringVertxReactBarAppApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(SpringVertxReactBarAppApplication.class, args);
-
-    }
-
-}
