--- src/main/java/io/github/mityavasilyev/springvertxreactbarapp/exceptions/DataNotFoundException.java
@@ -1,23 +1,0 @@
-package io.github.mityavasilyev.springvertxreactbarapp.exceptions;
-
-public class DataNotFoundException extends Exception{
-    /**
-     * Constructs a new exception with {@code null} as its detail message.
-     * The cause is not initialized, and may subsequently be initialized by a
-     * call to {@link #initCause}.
-     */
-    public DataNotFoundException() {
-    }
-
-    /**
-     * Constructs a new exception with the specified detail message.  The
-     * cause is not initialized, and may subsequently be initialized by
-     * a call to {@link #initCause}.
-     *
-     * @param message the detail message. The detail message is saved for
-     *                later retrieval by the {@link #getMessage()} method.
-     */
-    public DataNotFoundException(String message) {
-        super(message);
-    }
-}
