--- src/main/java/io/github/mityavasilyev/springvertxreactbarapp/extra/Recipe.java
@@ -1,33 +1,0 @@
-package io.github.mityavasilyev.springvertxreactbarapp.extra;
-
-import lombok.*;
-
-import javax.persistence.*;
-
-/**
- * Stores steps of the recipe for a cocktail
- */
-@Getter
-@Setter
-@AllArgsConstructor
-@NoArgsConstructor
-@Builder
-@Entity(name = "recipe")
-@Table(name = "recipes")
-public class Recipe {
-
-    @Id
-    @GeneratedValue(
-            strategy = GenerationType.IDENTITY
-    )
-    @Column(
-            updatable = false
-    )
-    private Long id;
-
-    @Column(
-            name = "steps",
-            columnDefinition = "TEXT"
-    )
-    private String steps;
-}
