--- src/main/java/io/github/mityavasilyev/springvertxreactbarapp/extra/Unit.java
@@ -1,8 +1,0 @@
-package io.github.mityavasilyev.springvertxreactbarapp.extra;
-
-/**
- * Provides unit type constants for measurements
- */
-public enum Unit {
-    OUNCE, MILLILITER, LITER, GRAM, PIECE
-}
