--- src/main/java/io/github/mityavasilyev/springvertxreactbarapp/product/Product.java
@@ -1,50 +1,0 @@
-package io.github.mityavasilyev.springvertxreactbarapp.product;
-
-import io.github.mityavasilyev.springvertxreactbarapp.extra.Unit;
-import lombok.*;
-
-import javax.persistence.*;
-
-/**
- * Describes consumable that will be used for ingredient
- */
-@Getter
-@Setter
-@AllArgsConstructor
-@NoArgsConstructor
-@Builder
-@Entity(name = "product")
-@Table(name = "products")
-public class Product {
-
-    @Id
-    @GeneratedValue(
-            strategy = GenerationType.IDENTITY
-    )
-    @Column(
-            updatable = false
-    )
-    private Long id;
-
-    @Column(
-            name = "name",
-            nullable = false,
-            columnDefinition = "TEXT"
-    )
-    private String name;
-
-    @Column(
-            name = "description",
-            columnDefinition = "TEXT"
-    )
-    private String description;
-
-    @Column(
-            name = "amount_left"
-    )
-    private Double amountLeft;
-
-    @Enumerated(EnumType.STRING)
-    private Unit unit;
-
-}
