+++ src/test/java/io/github/mityavasilyev/springreactbarapp/SpringReactBarAppApplicationTest.java
@@ -1,0 +1,16 @@
+package io.github.mityavasilyev.springreactbarapp;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+import static org.junit.jupiter.api.Assertions.*;
+
+@SpringBootTest
+class SpringReactBarAppApplicationTest {
+
+//    @Test
+//    void contextLoads() {
+//        assertEquals(true, true);
+//    }
+
+}
