--- src/test/java/io/github/mityavasilyev/springvertxreactbarapp/SpringVertxReactBarAppApplicationTests.java
@@ -1,13 +1,0 @@
-package io.github.mityavasilyev.springvertxreactbarapp;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class SpringVertxReactBarAppApplicationTests {
-
-    @Test
-    void contextLoads() {
-    }
-
-}
