--- src/test/java/io/github/mityavasilyev/springvertxreactbarapp/controllers/CocktailControllerTest.java
@@ -1,29 +1,0 @@
-package io.github.mityavasilyev.springvertxreactbarapp.controllers;
-
-import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.Test;
-
-import static org.junit.jupiter.api.Assertions.*;
-
-class CocktailControllerTest {
-
-    @BeforeEach
-    void setUp() {
-    }
-
-    @Test
-    void getAllCocktails() {
-    }
-
-    @Test
-    void getAllCocktailsByName() {
-    }
-
-    @Test
-    void getAllCocktailsByTagId() {
-    }
-
-    @Test
-    void getCocktailById() {
-    }
-}
